refactor: Model View 분리 작업
연산수 입력 메소드 분리
