+++ Calculator/Calculator/Model/CalculatorModel.swift
@@ -6,11 +6,41 @@
 //
 
 import Foundation
+import UIKit
 
 struct CalculatorModel {
     
+    var inputValue = ""
+    var presentValue = ""
+    var presentOperator = ""
+    var operatorStorage: [String] = []
+    var beforePresentValueStore: [String] = []
+    var userIsInTheAfterTabAnswerButton = false
+    var userIsInTheMiddleOfTyping = false
+    
+    mutating func addOperand(to buttonTitle: String) -> String {
+        if presentValue.contains(".") && buttonTitle == "." {
+            return presentValue
+        }
+        
+        if userIsInTheMiddleOfTyping {
+            inputValue += presentValue
+            inputValue += " \(operatorStorage.removeLast()) "
+            
+            presentValue = ""
+            operatorStorage = []
+            
+            presentValue += buttonTitle
+            return buttonTitle
+        }
+        presentValue += buttonTitle
+        userIsInTheMiddleOfTyping = false
+        return presentValue
+    }
+    
     mutating func checkHyphen(to presentValue: String) -> String {
         var checkHypenWord = presentValue
+        
         if checkHypenWord.contains("-") {
             let filterHypenWord = filterHypen(to: checkHypenWord)
             return filterHypenWord
@@ -21,6 +51,7 @@ struct CalculatorModel {
     
     private mutating func filterHypen(to checkHypenWord: String) -> String {
         var filterHypenWord = checkHypenWord
+        
         filterHypenWord = filterHypenWord.filter { (word: Character) in
             if word == "-" {
                 return false
